#544 Move Symbol classes into psyclone.psyir.symbols
