+++ src/psyclone/psyGen.py
@@ -49,6 +49,7 @@ from fparser.two import Fortran2003
 from psyclone.configuration import Config
 from psyclone.f2pygen import DirectiveGen
 from psyclone.core.access_info import VariablesAccessInfo, AccessType
+from psyclone.psyir.symbols import SymbolTable, SymbolError
 
 # We use the termcolor module (if available) to enable us to produce
 # coloured, textual representations of Invoke schedules. If it's not
@@ -251,21 +252,6 @@ class InternalError(Exception):
     def __init__(self, value):
         Exception.__init__(self, value)
         self.value = "PSyclone internal error: "+value
-
-    def __str__(self):
-        return str(self.value)
-
-
-class SymbolError(Exception):
-    '''
-    PSyclone-specific exception for use with errors relating to the SymbolTable
-    in the PSyIR.
-
-    :param str value: the message associated with the error.
-    '''
-    def __init__(self, value):
-        Exception.__init__(self, value)
-        self.value = "PSyclone SymbolTable error: "+value
 
     def __str__(self):
         return str(self.value)
@@ -5487,773 +5473,6 @@ class ACCDataDirective(ACCDirective):
         '''
         self._add_region(start_text="DATA", end_text="END DATA",
                          data_movement="analyse")
-
-
-@six.add_metaclass(abc.ABCMeta)
-class SymbolInterface(object):
-    '''
-    Abstract base class for capturing the access mechanism for symbols that
-    represent data that exists outside the section of code being represented
-    in the PSyIR.
-
-    :param access: How the symbol is accessed within the section of code or \
-                   None (if unknown).
-    :type access: :py:class:`psyclone.psyGen.DataSymbol.Access`
-    '''
-    def __init__(self, access=None):
-        self._access = None
-        # Use the setter as that has error checking
-        if not access:
-            self.access = DataSymbol.Access.UNKNOWN
-        else:
-            self.access = access
-
-    @property
-    def access(self):
-        '''
-        :returns: the access-type for this symbol.
-        :rtype: :py:class:`psyclone.psyGen.DataSymbol.Access`
-        '''
-        return self._access
-
-    @access.setter
-    def access(self, value):
-        '''
-        Setter for the access type of this symbol.
-
-        :param value: the new access type.
-        :type value: :py:class:`psyclon.psyGen.DataSymbol.Access`
-
-        :raises TypeError: if the supplied value is not of the correct type.
-        '''
-        if not isinstance(value, DataSymbol.Access):
-            raise TypeError(
-                "SymbolInterface.access must be a 'DataSymbol.Access' "
-                "but got '{0}'.".format(type(value)))
-        self._access = value
-
-
-class Symbol(object):
-    '''
-    Symbol item for the Symbol Table. It always has a fixed name label that
-    matches with the key on the SymbolTables that conatin the symbol.
-
-    :param str name: Name of the symbol.
-    '''
-
-    def __init__(self, name):
-        self._name = name
-
-    @property
-    def name(self):
-        '''
-        :returns: Name of the Symbol.
-        :rtype: str
-        '''
-        return self._name
-
-    def __str__(self):
-        return self._name
-
-
-class ContainerSymbol(Symbol):
-
-    def __init__(self, name):
-        super(ContainerSymbol, self).__init__(name)
-        self._reference = None
-
-    @property
-    def container(self):
-        if not self._reference:
-            self._import_container()
-        return self._reference
-
-    def _import_container(self):
-        # TODO: This is Fortran-specific, move to a sub-class or interface
-        from os import listdir, path
-        from fparser.two.parser import ParserFactory
-        from fparser.common.readfortran import FortranFileReader
-        from psyclone.psyir.frontend.fparser2 import Fparser2Reader
-
-        filename = self.name + '.f90'
-        for directory in Config.get().include_paths:
-            if filename in listdir(directory):
-                # Parse the module source code
-                reader = FortranFileReader(filename,
-                                           ignore_comments=True)
-                f2008_parser = ParserFactory().create(std="f2008")
-                ast = f2008_parser(reader)
-                fp2reader = Fparser2Reader()
-                self._reference = fp2reader.generate_container(ast)
-
-                # Finish the module search
-                break
-        else:
-            raise GenerationError(
-                "Module {0} not found in any of the include_path "
-                "directories {1}."
-                "".format(filename, Config.get().include_paths))
-
-    def __str__(self):
-        string = self._name + ": <"
-        if self._reference:
-            string += "linked>"
-        else:
-            string += "not linked>"
-        return string
-
-
-class DataSymbol(Symbol):
-    '''
-    Symbol identifying a data element. It contains information about:
-    the datatype, the shape (in column-major order) and, for a symbol
-    representing data that exists outside of the local scope, the interface
-    to that symbol (i.e. the mechanism by which it is accessed).
-
-    :param str name: Name of the symbol.
-    :param str datatype: Data type of the symbol. (One of \
-                     :py:attr:`psyclone.psyGen.DataSymbol.valid_data_types`.)
-    :param list shape: Shape of the symbol in column-major order (leftmost \
-                       index is contiguous in memory). Each entry represents \
-                       an array dimension. If it is 'None' the extent of that \
-                       dimension is unknown, otherwise it holds an integer \
-                       literal or a reference to an integer symbol with the \
-                       extent. If it is an empty list then the symbol \
-                       represents a scalar.
-    :param interface: Object describing the interface to this symbol (i.e. \
-                      whether it is passed as a routine argument or accessed \
-                      in some other way) or None if the symbol is local.
-    :type interface: :py:class:`psyclone.psyGen.SymbolInterface` or NoneType.
-    :param constant_value: Sets a fixed known value for this DataSymbol. \
-                           If the value is None (the default) \
-                           then this symbol is not a constant. The \
-                           datatype of the constant value must be \
-                           compatible with the datatype of the symbol.
-    :type constant_value: int, str or bool
-
-    :raises NotImplementedError: Provided parameters are not supported yet.
-    :raises TypeError: Provided parameters have invalid error type.
-    :raises ValueError: Provided parameters contain invalid values.
-
-    '''
-    ## Tuple with the valid datatypes.
-    valid_data_types = ('real',  # Floating point
-                        'integer',
-                        'character',
-                        'boolean',
-                        'deferred')  # Type of this symbol not yet determined
-
-    ## Mapping from supported data types for constant values to
-    #  internal Python types
-    mapping = {'integer': int,
-               'character': str,
-               'boolean': bool,
-               'real': float}
-
-    class Access(Enum):
-        '''
-        Enumeration for the different types of access that a DataSymbol is
-        permitted to have.
-
-        '''
-        ## The symbol is only ever read within the current scoping block.
-        READ = 1
-        ## The first access of the symbol in the scoping block is a write and
-        # therefore any value that it may have had upon entry is discarded.
-        WRITE = 2
-        ## The first access of the symbol in the scoping block is a read but
-        # it is subsequently written to.
-        READWRITE = 3
-        ## The way in which the symbol is accessed in the scoping block is
-        # unknown
-        UNKNOWN = 4
-
-    class Argument(SymbolInterface):
-        '''
-        Captures the interface to a symbol that is accessed as a routine
-        argument.
-
-        :param access: how the symbol is accessed within the local scope.
-        :type access: :py:class:`psyclone.psyGen.DataSymbol.Access`
-        '''
-        def __init__(self, access=None):
-            super(DataSymbol.Argument, self).__init__(access=access)
-            self._pass_by_value = False
-
-        def __str__(self):
-            return "Argument(pass-by-value={0})".format(self._pass_by_value)
-
-    class FortranGlobal(SymbolInterface):
-        '''
-        Describes the interface to a Fortran DataSymbol representing data that
-        is supplied as some sort of global variable. Currently only supports
-        data accessed via a module 'USE' statement.
-
-        :param container_symbol: symbol of the external container from which \
-                                 the symbol is imported.
-        :type container_symbol: :py:class:`psyclone.psyGen.ContainerSymbol`
-        :param access: the manner in which the DataSymbol is accessed in the \
-                       associated code section. If None is supplied then the \
-                       access is DataSymbol.Access.UNKNOWN.
-        :type access: :py:class:`psyclone.psyGen.DataSymbol.Access` or None.
-
-        :raise TypeError: if the container_symbol is not a ContainerSymbol
-        '''
-        def __init__(self, container_symbol, access=None):
-            super(DataSymbol.FortranGlobal, self).__init__(access=access)
-
-            if not isinstance(container_symbol, ContainerSymbol):
-                raise TypeError(
-                    "FortranGlobal container_symbol parameter must be of type"
-                    " ContainerSymbol, but found {0}."
-                    "".format(type(container_symbol)))
-
-            self._container_symbol = container_symbol
-
-        @property
-        def container_symbol(self):
-            return self._container_symbol
-
-        def __str__(self):
-            return "FortranModule({0})".format(self.container_symbol.name)
-
-    def __init__(self, name, datatype, shape=None, constant_value=None,
-                 interface=None):
-
-        super(DataSymbol, self).__init__(name)
-
-        if datatype not in DataSymbol.valid_data_types:
-            raise NotImplementedError(
-                "DataSymbol can only be initialised with {0} datatypes but "
-                "found '{1}'.".format(str(DataSymbol.valid_data_types),
-                                      datatype))
-        self._datatype = datatype
-
-        if shape is None:
-            shape = []
-        elif not isinstance(shape, list):
-            raise TypeError("DataSymbol shape attribute must be a list.")
-
-        for dimension in shape:
-            if isinstance(dimension, DataSymbol):
-                if dimension.datatype != "integer" or dimension.shape:
-                    raise TypeError(
-                        "DataSymbols that are part of another symbol shape can"
-                        " only be scalar integers, but found '{0}'."
-                        "".format(str(dimension)))
-            elif not isinstance(dimension, (type(None), int)):
-                raise TypeError("DataSymbol shape list elements can only be "
-                                "'DataSymbol', 'integer' or 'None'.")
-        self._shape = shape
-        # The following attributes have setter methods (with error checking)
-        self._constant_value = None
-        self._interface = None
-        # If an interface is specified for this symbol then the data with
-        # which it is associated must exist outside of this kernel.
-        self.interface = interface
-        self.constant_value = constant_value
-
-    def resolve_deferred(self):
-        if self.datatype == "deferred":
-            if isinstance(self.interface, DataSymbol.FortranGlobal):
-                # Copy all the symbol properties but the interface
-                tmp = self.interface
-                module = self.interface.container_symbol
-                extern_symbol = module.container.symbol_table.lookup(self.name)
-                self.copy_properties(extern_symbol)
-                self.interface = tmp
-            else:
-                raise NotImplementedError(
-                    "Lazy evalution of deferred {0} is not supported yet."
-                    "".format(self.interface))
-
-    @property
-    def datatype(self):
-        '''
-        :returns: Datatype of the DataSymbol.
-        :rtype: str
-        '''
-        return self._datatype
-
-    @property
-    def access(self):
-        '''
-        :returns: How this symbol is accessed (read, readwrite etc.) within \
-                  the local scope.
-        :rtype: :py:class:`psyclone.psyGen.DataSymbol.Access` or NoneType.
-        '''
-        if self._interface:
-            return self._interface.access
-        # This symbol has no interface info and therefore is local
-        return None
-
-    @property
-    def shape(self):
-        '''
-        :returns: Shape of the symbol in column-major order (leftmost \
-                  index is contiguous in memory). Each entry represents \
-                  an array dimension. If it is 'None' the extent of that \
-                  dimension is unknown, otherwise it holds an integer \
-                  literal or a reference to an integer symbol with the \
-                  extent. If it is an empty list then the symbol \
-                  represents a scalar.
-        :rtype: list
-        '''
-        return self._shape
-
-    @property
-    def scope(self):
-        '''
-        :returns: Whether the symbol is 'local' (just exists inside the \
-                  kernel scope) or 'global' (data also lives outside the \
-                  kernel). Global-scoped symbols must have an associated \
-                  'interface' that specifies the mechanism by which the \
-                  kernel accesses the associated data.
-        :rtype: str
-        '''
-        if self._interface:
-            return "global"
-        return "local"
-
-    @property
-    def interface(self):
-        '''
-        :returns: the an object describing the external interface to \
-                  this DataSymbol or None (if it is local).
-        :rtype: Sub-class of :py:class:`psyclone.psyGen.SymbolInterface` or \
-                NoneType.
-        '''
-        return self._interface
-
-    @interface.setter
-    def interface(self, value):
-        '''
-        Setter for the Interface associated with this DataSymbol.
-
-        :param value: an Interface object describing how the DataSymbol is \
-                      accessed by the code or None if it is local.
-        :type value: Sub-class of :py:class:`psyclone.psyGen.SymbolInterface` \
-                     or NoneType.
-
-        :raises TypeError: if the supplied `value` is of the wrong type.
-        '''
-        if value is not None and not isinstance(value, SymbolInterface):
-            raise TypeError("The interface to a DataSymbol must be a "
-                            "SymbolInterface or None but got '{0}'".
-                            format(type(value)))
-        self._interface = value
-
-    @property
-    def is_constant(self):
-        '''
-        :returns: Whether the symbol is a constant with a fixed known \
-        value (True) or not (False).
-        :rtype: bool
-
-        '''
-        return self._constant_value is not None
-
-    @property
-    def is_scalar(self):
-        '''
-        :returns: True if this symbol is a scalar and False otherwise.
-        :rtype: bool
-
-        '''
-        # If the shape variable is an empty list then this symbol is a
-        # scalar.
-        return self.shape == []
-
-    @property
-    def is_array(self):
-        '''
-        :returns: True if this symbol is an array and False otherwise.
-        :rtype: bool
-
-        '''
-        # The assumption in this method is that if this symbol is not
-        # a scalar then it is an array. If this assumption becomes
-        # invalid then this logic will need to be changed
-        # appropriately.
-        return not self.is_scalar
-
-    @property
-    def constant_value(self):
-        '''
-        :returns: The fixed known value for this symbol if one has \
-        been set or None if not.
-        :rtype: int, str, bool or NoneType
-
-        '''
-        return self._constant_value
-
-    @constant_value.setter
-    def constant_value(self, new_value):
-        '''
-        :param constant_value: Set or change the fixed known value of \
-        the constant for this DataSymbol. If the value is None then this \
-        symbol does not have a fixed constant. The datatype of \
-        new_value must be compatible with the datatype of the symbol.
-        :type constant_value: int, str or bool
-
-        :raises ValueError: If a non-None value is provided and 1) \
-        this DataSymbol instance does not have local scope, or 2) this \
-        DataSymbol instance is not a scalar (as the shape attribute is not \
-        empty), or 3) a constant value is provided but the type of the \
-        value does not support this, or 4) the type of the value \
-        provided is not compatible with the datatype of this DataSymbol \
-        instance.
-
-        '''
-        if new_value is not None:
-            if self.scope != "local":
-                raise ValueError(
-                    "DataSymbol with a constant value is currently limited to "
-                    "having local scope but found '{0}'.".format(self.scope))
-            if self.is_array:
-                raise ValueError(
-                    "DataSymbol with a constant value must be a scalar but the"
-                    " shape attribute is not empty.")
-            try:
-                lookup = DataSymbol.mapping[self.datatype]
-            except KeyError:
-                raise ValueError(
-                    "A constant value is not supported for "
-                    "datatype '{0}'.".format(self.datatype))
-            if not isinstance(new_value, lookup):
-                raise ValueError(
-                    "This DataSymbol instance's datatype is '{0}' which means "
-                    "the constant value is expected to be '{1}' but found "
-                    "'{2}'.".format(self.datatype,
-                                    DataSymbol.mapping[self.datatype],
-                                    type(new_value)))
-        self._constant_value = new_value
-
-    def __str__(self):
-        ret = self.name + ": <" + self.datatype + ", "
-        if self.is_array:
-            ret += "Array["
-            for dimension in self.shape:
-                if isinstance(dimension, DataSymbol):
-                    ret += dimension.name
-                elif isinstance(dimension, int):
-                    ret += str(dimension)
-                elif dimension is None:
-                    ret += "'Unknown bound'"
-                else:
-                    raise InternalError(
-                        "DataSymbol shape list elements can only be "
-                        "'DataSymbol', 'integer' or 'None', but found '{0}'."
-                        "".format(type(dimension)))
-                ret += ", "
-            ret = ret[:-2] + "]"  # Deletes last ", " and adds "]"
-        else:
-            ret += "Scalar"
-        if self.interface:
-            ret += ", global=" + str(self.interface)
-        else:
-            ret += ", local"
-        if self.is_constant:
-            ret += ", constant_value={0}".format(self.constant_value)
-        return ret + ">"
-
-    def copy(self):
-        '''Create and return a copy of this object. Any references to the
-        original will not be affected so the copy will not be referred
-        to by any other object.
-
-        :returns: A symbol object with the same properties as this \
-                  symbol object.
-        :rtype: :py:class:`psyclone.psyGen.DataSymbol`
-
-        '''
-        return DataSymbol(self.name, self.datatype, shape=self.shape[:],
-                          constant_value=self.constant_value,
-                          interface=self.interface)
-
-    def copy_properties(self, symbol_in):
-        '''Replace all properties in this object with the properties from
-        symbol_in, apart from the name which is immutable.
-
-        :param symbol_in: The symbol from which the properties are \
-                          copied from.
-        :type symbol_in: :py:class:`psyclone.psyGen.DataSymbol`
-
-        :raises TypeError: If the argument is not the expected type.
-
-        '''
-        if not isinstance(symbol_in, DataSymbol):
-            raise TypeError("Argument should be of type 'DataSymbol' but found"
-                            " '{0}'.".format(type(symbol_in).__name__))
-
-        self._datatype = symbol_in.datatype
-        self._shape = symbol_in.shape[:]
-        self._constant_value = symbol_in.constant_value
-        self._interface = symbol_in.interface
-
-
-class SymbolTable(object):
-    '''
-    Encapsulates the symbol table and provides methods to add new symbols
-    and look up existing symbols. It is implemented as a single scope
-    symbol table (nested scopes not supported).
-
-    :param kernel: Reference to the KernelSchedule to which this symbol table \
-        belongs.
-    :type kernel: :py:class:`psyclone.psyGen.KernelSchedule` or NoneType
-    '''
-    # TODO: (Issue #321) Explore how the SymbolTable overlaps with the
-    # NameSpace class functionality.
-    def __init__(self, kernel=None):
-        # Dict of Symbol objects with the symbol names as keys. Make
-        # this ordered so that different versions of Python always
-        # produce code with declarations in the same order.
-        self._symbols = OrderedDict()
-        # Ordered list of the arguments.
-        self._argument_list = []
-        # Reference to KernelSchedule to which this symbol table belongs.
-        self._kernel = kernel
-
-    def add(self, new_symbol):
-        '''Add a new symbol to the symbol table.
-
-        :param new_symbol: The symbol to add to the symbol table.
-        :type new_symbol: :py:class:`psyclone.psyGen.Symbol`
-
-        :raises KeyError: If the symbol name is already in use.
-
-        '''
-        if new_symbol.name in self._symbols:
-            raise KeyError("Symbol table already contains a symbol with"
-                           " name '{0}'.".format(new_symbol.name))
-        self._symbols[new_symbol.name] = new_symbol
-
-    def swap_symbol_properties(self, symbol1, symbol2):
-        '''Swaps the properties of symbol1 and symbol2 apart from the symbol
-        name. Argument list positions are also updated appropriately.
-
-        :param symbol1: The first symbol.
-        :type symbol1: :py:class:`psyclone.psyGen.Symbol`
-        :param symbol2: The second symbol.
-        :type symbol2: :py:class:`psyclone.psyGen.Symbol`
-
-        :raises KeyError: If either of the supplied symbols are not in \
-                          the symbol table.
-        :raises TypeError: If the supplied arguments are not symbols, \
-                 or the names of the symbols are the same in the SymbolTable \
-                 instance.
-
-        '''
-        for symbol in [symbol1, symbol2]:
-            if not isinstance(symbol, Symbol):
-                raise TypeError("Arguments should be of type 'Symbol' but "
-                                "found '{0}'.".format(type(symbol).__name__))
-            if symbol.name not in self._symbols:
-                raise KeyError("Symbol '{0}' is not in the symbol table."
-                               "".format(symbol.name))
-        if symbol1.name == symbol2.name:
-            raise ValueError("The symbols should have different names, but "
-                             "found '{0}' for both.".format(symbol1.name))
-
-        tmp_symbol = symbol1.copy()
-        symbol1.copy_properties(symbol2)
-        symbol2.copy_properties(tmp_symbol)
-
-        # Update argument list if necessary
-        index1 = None
-        if symbol1 in self._argument_list:
-            index1 = self._argument_list.index(symbol1)
-        index2 = None
-        if symbol2 in self._argument_list:
-            index2 = self._argument_list.index(symbol2)
-        if index1 is not None:
-            self._argument_list[index1] = symbol2
-        if index2 is not None:
-            self._argument_list[index2] = symbol1
-
-    def specify_argument_list(self, argument_symbols):
-        '''
-        Sets-up the internal list storing the order of the arguments to this
-        kernel.
-
-        :param list argument_symbols: Ordered list of the DataSymbols \
-                                      representing the kernel arguments.
-
-        :raises ValueError: If the new argument_list is not consistent with \
-                            the existing entries in the SymbolTable.
-
-        '''
-        self._validate_arg_list(argument_symbols)
-        self._argument_list = argument_symbols[:]
-
-    def lookup(self, name):
-        '''
-        Look up a symbol in the symbol table.
-
-        :param str name: Name of the symbol
-        :raises KeyError: If the given name is not in the Symbol Table.
-
-        '''
-        try:
-            return self._symbols[name]
-        except KeyError:
-            raise KeyError("Could not find '{0}' in the Symbol Table."
-                           "".format(name))
-
-    def __contains__(self, key):
-        '''Check if the given key is part of the Symbol Table.
-
-        :param str key: key to check for existance.
-        :returns: Whether the Symbol Table contains the given key.
-        :rtype: bool
-        '''
-        return key in self._symbols
-
-    @property
-    def argument_list(self):
-        '''
-        Checks that the contents of the SymbolTable are self-consistent
-        and then returns the list of kernel arguments.
-
-        :returns: Ordered list of arguments.
-        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
-
-        :raises InternalError: If the entries of the SymbolTable are not \
-                               self-consistent.
-
-        '''
-        try:
-            self._validate_arg_list(self._argument_list)
-            self._validate_non_args()
-        except ValueError as err:
-            # If the SymbolTable is inconsistent at this point then
-            # we have an InternalError.
-            raise InternalError(str(err.args))
-        return self._argument_list
-
-    @staticmethod
-    def _validate_arg_list(arg_list):
-        '''
-        Checks that the supplied list of Symbols are valid kernel arguments.
-
-        :param arg_list: the proposed kernel arguments.
-        :type param_list: list of :py:class:`psyclone.psyGen.DataSymbol`
-
-        :raises TypeError: if any item in the supplied list is not a \
-            DataSymbol.
-        :raises ValueError: if any of the symbols has no Interface.
-        :raises ValueError: if any of the symbols has an Interface and is not \
-                            a :py:class:`psyclone.psyGen.DataSymbol.Argument`.
-
-        '''
-        for symbol in arg_list:
-            if not isinstance(symbol, DataSymbol):
-                raise TypeError("Expected a list of DataSymbols but found an "
-                                "object of type '{0}'.".format(type(symbol)))
-            # All symbols in the argument list must have a
-            # 'Symbol.Argument' interface
-            if symbol.scope == 'local':
-                raise ValueError(
-                    "DataSymbol '{0}' is listed as a kernel argument but has "
-                    "no associated Interface.".format(str(symbol)))
-            if not isinstance(symbol.interface, DataSymbol.Argument):
-                raise ValueError(
-                    "DataSymbol '{0}' is listed as a kernel argument but has "
-                    "an interface of type '{1}' rather than "
-                    "DataSymbol.Argument"
-                    "".format(str(symbol), type(symbol.interface)))
-
-    def _validate_non_args(self):
-        '''
-        Performs internal consistency checks on the current entries in the
-        SymbolTable that do not represent kernel arguments.
-
-        :raises ValueError: If a symbol that is not in the argument list \
-                            has a Symbol.Argument interface.
-
-        '''
-        for symbol in self.variables:
-            if symbol not in self._argument_list:
-                # DataSymbols not in the argument list must not have a
-                # Symbol.Argument interface
-                if symbol.interface and isinstance(symbol.interface,
-                                                   DataSymbol.Argument):
-                    raise ValueError(
-                        "Symbol '{0}' is not listed as a kernel argument and "
-                        "yet has a DataSymbol.Argument interface.".format(
-                            str(symbol)))
-
-    @property
-    def symbols(self):
-        '''
-        :returns:  List of symbols.
-        :rtype: list of :py:class:`psyclone.psyGen.Symbol`
-        '''
-        return list(self._symbols.values())
-
-    @property
-    def variables(self):
-        '''
-        :returns:  List of symbols representing data variables.
-        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
-        '''
-        return [sym for sym in self._symbols.values() if
-                isinstance(sym, DataSymbol)]
-
-    @property
-    def local_variables(self):
-        '''
-        :returns:  List of symbols representing local variables.
-        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
-        '''
-        return [sym for sym in self.variables if sym.scope == "local"]
-
-    @property
-    def global_variables(self):
-        '''
-        :returns: list of symbols that are not routine arguments but \
-                  still have 'global' scope - i.e. are associated with \
-                  data that exists outside the current scope.
-        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
-
-        '''
-        return [sym for sym in self.variables if sym.scope == "global"
-                and not isinstance(sym.interface, DataSymbol.Argument)]
-
-    @property
-    def iteration_indices(self):
-        '''
-        :returns: List of symbols representing kernel iteration indices.
-        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
-
-        :raises NotImplementedError: this method is abstract.
-        '''
-        raise NotImplementedError(
-            "Abstract property. Which symbols are iteration indices is"
-            " API-specific.")
-
-    @property
-    def data_arguments(self):
-        '''
-        :returns: List of symbols representing kernel data arguments.
-        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
-
-        :raises NotImplementedError: this method is abstract.
-        '''
-        raise NotImplementedError(
-            "Abstract property. Which symbols are data arguments is"
-            " API-specific.")
-
-    def view(self):
-        '''
-        Print a representation of this Symbol Table to stdout.
-        '''
-        print(str(self))
-
-    def __str__(self):
-        return ("Symbol Table:\n" +
-                "\n".join(map(str, self._symbols.values())) +
-                "\n")
 
 
 class KernelSchedule(Schedule):
