+++ src/psyclone/psyir/symbols/SymbolTable.py
@@ -1,0 +1,307 @@
+# -----------------------------------------------------------------------------
+# BSD 3-Clause License
+#
+# Copyright (c) 2017-2019, Science and Technology Facilities Council.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# * Redistributions of source code must retain the above copyright notice, this
+#   list of conditions and the following disclaimer.
+#
+# * Redistributions in binary form must reproduce the above copyright notice,
+#   this list of conditions and the following disclaimer in the documentation
+#   and/or other materials provided with the distribution.
+#
+# * Neither the name of the copyright holder nor the names of its
+#   contributors may be used to endorse or promote products derived from
+#   this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+# -----------------------------------------------------------------------------
+# Authors R. W. Ford, A. R. Porter and S. Siso, STFC Daresbury Lab
+#         I. Kavcic, Met Office
+#         J. Henrichs, Bureau of Meteorology
+# -----------------------------------------------------------------------------
+
+''' File Description '''
+
+from collections import OrderedDict
+from psyclone.psyir.symbols import Symbol, DataSymbol
+
+
+class SymbolTable(object):
+    '''
+    Encapsulates the symbol table and provides methods to add new symbols
+    and look up existing symbols. It is implemented as a single scope
+    symbol table (nested scopes not supported).
+
+    :param kernel: Reference to the KernelSchedule to which this symbol table \
+        belongs.
+    :type kernel: :py:class:`psyclone.psyGen.KernelSchedule` or NoneType
+    '''
+    # TODO: (Issue #321) Explore how the SymbolTable overlaps with the
+    # NameSpace class functionality.
+    def __init__(self, kernel=None):
+        # Dict of Symbol objects with the symbol names as keys. Make
+        # this ordered so that different versions of Python always
+        # produce code with declarations in the same order.
+        self._symbols = OrderedDict()
+        # Ordered list of the arguments.
+        self._argument_list = []
+        # Reference to KernelSchedule to which this symbol table belongs.
+        self._kernel = kernel
+
+    def add(self, new_symbol):
+        '''Add a new symbol to the symbol table.
+
+        :param new_symbol: The symbol to add to the symbol table.
+        :type new_symbol: :py:class:`psyclone.psyGen.Symbol`
+
+        :raises KeyError: If the symbol name is already in use.
+
+        '''
+        if new_symbol.name in self._symbols:
+            raise KeyError("Symbol table already contains a symbol with"
+                           " name '{0}'.".format(new_symbol.name))
+        self._symbols[new_symbol.name] = new_symbol
+
+    def swap_symbol_properties(self, symbol1, symbol2):
+        '''Swaps the properties of symbol1 and symbol2 apart from the symbol
+        name. Argument list positions are also updated appropriately.
+
+        :param symbol1: The first symbol.
+        :type symbol1: :py:class:`psyclone.psyGen.Symbol`
+        :param symbol2: The second symbol.
+        :type symbol2: :py:class:`psyclone.psyGen.Symbol`
+
+        :raises KeyError: If either of the supplied symbols are not in \
+                          the symbol table.
+        :raises TypeError: If the supplied arguments are not symbols, \
+                 or the names of the symbols are the same in the SymbolTable \
+                 instance.
+
+        '''
+        for symbol in [symbol1, symbol2]:
+            if not isinstance(symbol, Symbol):
+                raise TypeError("Arguments should be of type 'Symbol' but "
+                                "found '{0}'.".format(type(symbol).__name__))
+            if symbol.name not in self._symbols:
+                raise KeyError("Symbol '{0}' is not in the symbol table."
+                               "".format(symbol.name))
+        if symbol1.name == symbol2.name:
+            raise ValueError("The symbols should have different names, but "
+                             "found '{0}' for both.".format(symbol1.name))
+
+        tmp_symbol = symbol1.copy()
+        symbol1.copy_properties(symbol2)
+        symbol2.copy_properties(tmp_symbol)
+
+        # Update argument list if necessary
+        index1 = None
+        if symbol1 in self._argument_list:
+            index1 = self._argument_list.index(symbol1)
+        index2 = None
+        if symbol2 in self._argument_list:
+            index2 = self._argument_list.index(symbol2)
+        if index1 is not None:
+            self._argument_list[index1] = symbol2
+        if index2 is not None:
+            self._argument_list[index2] = symbol1
+
+    def specify_argument_list(self, argument_symbols):
+        '''
+        Sets-up the internal list storing the order of the arguments to this
+        kernel.
+
+        :param list argument_symbols: Ordered list of the DataSymbols \
+                                      representing the kernel arguments.
+
+        :raises ValueError: If the new argument_list is not consistent with \
+                            the existing entries in the SymbolTable.
+
+        '''
+        self._validate_arg_list(argument_symbols)
+        self._argument_list = argument_symbols[:]
+
+    def lookup(self, name):
+        '''
+        Look up a symbol in the symbol table.
+
+        :param str name: Name of the symbol
+        :raises KeyError: If the given name is not in the Symbol Table.
+
+        '''
+        try:
+            return self._symbols[name]
+        except KeyError:
+            raise KeyError("Could not find '{0}' in the Symbol Table."
+                           "".format(name))
+
+    def __contains__(self, key):
+        '''Check if the given key is part of the Symbol Table.
+
+        :param str key: key to check for existance.
+        :returns: Whether the Symbol Table contains the given key.
+        :rtype: bool
+        '''
+        return key in self._symbols
+
+    @property
+    def argument_list(self):
+        '''
+        Checks that the contents of the SymbolTable are self-consistent
+        and then returns the list of kernel arguments.
+
+        :returns: Ordered list of arguments.
+        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
+
+        :raises InternalError: If the entries of the SymbolTable are not \
+                               self-consistent.
+
+        '''
+        from psyclone.psyGen import InternalError
+        try:
+            self._validate_arg_list(self._argument_list)
+            self._validate_non_args()
+        except ValueError as err:
+            # If the SymbolTable is inconsistent at this point then
+            # we have an InternalError.
+            raise InternalError(str(err.args))
+        return self._argument_list
+
+    @staticmethod
+    def _validate_arg_list(arg_list):
+        '''
+        Checks that the supplied list of Symbols are valid kernel arguments.
+
+        :param arg_list: the proposed kernel arguments.
+        :type param_list: list of :py:class:`psyclone.psyGen.DataSymbol`
+
+        :raises TypeError: if any item in the supplied list is not a \
+            DataSymbol.
+        :raises ValueError: if any of the symbols has no Interface.
+        :raises ValueError: if any of the symbols has an Interface and is not \
+                            a :py:class:`psyclone.psyGen.DataSymbol.Argument`.
+
+        '''
+        for symbol in arg_list:
+            if not isinstance(symbol, DataSymbol):
+                raise TypeError("Expected a list of DataSymbols but found an "
+                                "object of type '{0}'.".format(type(symbol)))
+            # All symbols in the argument list must have a
+            # 'Symbol.Argument' interface
+            if symbol.scope == 'local':
+                raise ValueError(
+                    "DataSymbol '{0}' is listed as a kernel argument but has "
+                    "no associated Interface.".format(str(symbol)))
+            if not isinstance(symbol.interface, DataSymbol.Argument):
+                raise ValueError(
+                    "DataSymbol '{0}' is listed as a kernel argument but has "
+                    "an interface of type '{1}' rather than "
+                    "DataSymbol.Argument"
+                    "".format(str(symbol), type(symbol.interface)))
+
+    def _validate_non_args(self):
+        '''
+        Performs internal consistency checks on the current entries in the
+        SymbolTable that do not represent kernel arguments.
+
+        :raises ValueError: If a symbol that is not in the argument list \
+                            has a Symbol.Argument interface.
+
+        '''
+        for symbol in self.variables:
+            if symbol not in self._argument_list:
+                # DataSymbols not in the argument list must not have a
+                # Symbol.Argument interface
+                if symbol.interface and isinstance(symbol.interface,
+                                                   DataSymbol.Argument):
+                    raise ValueError(
+                        "Symbol '{0}' is not listed as a kernel argument and "
+                        "yet has a DataSymbol.Argument interface.".format(
+                            str(symbol)))
+
+    @property
+    def symbols(self):
+        '''
+        :returns:  List of symbols.
+        :rtype: list of :py:class:`psyclone.psyGen.Symbol`
+        '''
+        return list(self._symbols.values())
+
+    @property
+    def variables(self):
+        '''
+        :returns:  List of symbols representing data variables.
+        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
+        '''
+        return [sym for sym in self._symbols.values() if
+                isinstance(sym, DataSymbol)]
+
+    @property
+    def local_variables(self):
+        '''
+        :returns:  List of symbols representing local variables.
+        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
+        '''
+        return [sym for sym in self.variables if sym.scope == "local"]
+
+    @property
+    def global_variables(self):
+        '''
+        :returns: list of symbols that are not routine arguments but \
+                  still have 'global' scope - i.e. are associated with \
+                  data that exists outside the current scope.
+        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
+
+        '''
+        return [sym for sym in self.variables if sym.scope == "global"
+                and not isinstance(sym.interface, DataSymbol.Argument)]
+
+    @property
+    def iteration_indices(self):
+        '''
+        :returns: List of symbols representing kernel iteration indices.
+        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
+
+        :raises NotImplementedError: this method is abstract.
+        '''
+        raise NotImplementedError(
+            "Abstract property. Which symbols are iteration indices is"
+            " API-specific.")
+
+    @property
+    def data_arguments(self):
+        '''
+        :returns: List of symbols representing kernel data arguments.
+        :rtype: list of :py:class:`psyclone.psyGen.DataSymbol`
+
+        :raises NotImplementedError: this method is abstract.
+        '''
+        raise NotImplementedError(
+            "Abstract property. Which symbols are data arguments is"
+            " API-specific.")
+
+    def view(self):
+        '''
+        Print a representation of this Symbol Table to stdout.
+        '''
+        print(str(self))
+
+    def __str__(self):
+        return ("Symbol Table:\n" +
+                "\n".join(map(str, self._symbols.values())) +
+                "\n")
