+++ src/psyclone/psyir/symbols/__init__.py
@@ -1,0 +1,11 @@
+
+from psyclone.psyir.symbols.Symbol import Symbol, SymbolError
+from psyclone.psyir.symbols.DataSymbol import DataSymbol
+from psyclone.psyir.symbols.ContainerSymbol import ContainerSymbol
+from psyclone.psyir.symbols.SymbolTable import SymbolTable
+
+__all__ = ['Symbol',
+           'SymbolError',
+           'SymbolTable',
+           'DataSymbol',
+           'ContainerSymbol']
