+++ src/psyclone/tests/gocean1p0_opencl_test.py
@@ -43,7 +43,8 @@ import pytest
 from psyclone.configuration import Config
 from psyclone.transformations import OCLTrans
 from psyclone.gocean1p0 import GOKernelSchedule
-from psyclone.psyGen import GenerationError, DataSymbol
+from psyclone.psyGen import GenerationError
+from psyclone.psyir.symbols import DataSymbol
 from psyclone.tests.utilities import Compile, get_invoke
 
 from psyclone.tests.gocean1p0_build import GOcean1p0OpenCLBuild
