+++ src/psyclone/tests/psyir/backend/c_test.py
@@ -43,9 +43,10 @@ import pytest
 
 from psyclone.psyir.backend.visitor import VisitorError
 from psyclone.psyir.backend.c import CWriter
-from psyclone.psyGen import DataSymbol, Node, CodeBlock, Assignment, \
+from psyclone.psyGen import Node, CodeBlock, Assignment, \
     Reference, Return, Array, Literal, UnaryOperation, BinaryOperation, \
     Schedule
+from psyclone.psyir.symbols import DataSymbol
 
 
 def test_cw_gen_declaration():
